[SYNC-210][FE] refactor: Split login and signup modals.
