+++ src/components/Account/PrivacyPolicyModal.vue
